requirement.txt
absl-py==0.7.1
apipkg==1.5
asn1crypto==0.22.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcolz==1.2.1
bhtsne==0.1.9
bleach==3.1.0
bokeh==1.2.0
certifi==2019.3.9
cffi==1.10.0
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
cnn-finetune==0.6.0
coverage==4.5.3
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.9.0.1
dask==1.2.2
decorator==4.4.0
defusedxml==0.6.0
distributed==1.28.1
entrypoints==0.3
execnet==1.6.0
flaky==3.5.3
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
heapdict==1.0.0
idna==2.6
importlib-metadata==0.18
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
-e git://github.com/keras-team/keras.git@cd3f27b5d2ccd843a2c98ee196f81217542e36c8#egg=Keras
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
locket==0.2.0
Mako==1.0.10
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.13
mkl-random==1.0.4
mkl-service==2.0.2
mock==3.0.5
more-itertools==7.0.0
msgpack==0.6.1
munch==2.3.2
nbconvert==5.5.0
nbformat==4.4.0
nmslib==1.8.1
nose==1.3.7
notebook==5.7.8
numexpr==2.6.9
numpy==1.16.4
olefile==0.46
packaging==16.8
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.12.0
pretrainedmodels==0.7.4
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.2
ptyprocess==0.6.0
py==1.8.0
pybind11==2.4.2
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
Pygments==2.4.2
pygpu==0.7.6
pyOpenSSL==17.0.0
pyparsing==2.2.0
pyrsistent==0.14.11
PySocks==1.7.0
pytest==4.6.3
pytest-cache==1.0
pytest-cov==2.7.1
pytest-forked==1.0.2
pytest-pep8==1.0.6
pytest-xdist==1.29.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.0
requests==2.22.0
ruamel-yaml==0.15.46
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
six==1.12.0
sklearn-pandas==1.8.0
sortedcontainers==2.1.0
tables==3.5.2
tblib==1.4.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.2
tqdm==4.36.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
zict==0.1.4
zipp==0.5.1
